+++ sw/blas/gemm/src/gemm_occamy_2dpipe.h
@@ -237,7 +237,7 @@ void gemm_oc_opt2d(double alpha, double beta, uint32_t m, uint32_t n,
 inline void gemm_oc(precision_t prec, uint32_t expand, uint32_t setup_ssr,
                     uint32_t transa, uint32_t transb, uint32_t m, uint32_t n,
                     uint32_t k, double alpha, void* a, uint32_t lda, void* b,
-                    uint32_t ldb, uint32_t beta, void* c, uint32_t ldc) {
+                    uint32_t ldb, double beta, void* c, uint32_t ldc) {
     // gemm_cluster_kernel(alpha, beta, m, n, k, a, b, c, lda, ldb, ldc);
     // snrt_fpu_fence();
     // snrt_cluster_hw_barrier();
